Change reducer to “sector”
Just a proposal.  Been thinking about the ‘addInterfaceReducer’ name and although they by definition are reducers I feel perhaps ‘selector’ is a better name.  The job of a sector is to receive a set of keys and say “I know the implementation you need, it’s this key”.  A selector can either pass the keys or a subset of keys to the next selector OR “select” the answer by returning a single key.  

@josheverett